requirements.txt
colorama>=0.4.4
openai
python-dotenv>=0.19.1
rdflib
tiktoken
pyyaml